api-server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
alembic==1.10.4
python-dotenv==1.0.0
boto3==1.26.133
ffmpeg-python==0.2.0 